+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/application/feign/UserService.java
@@ -2,7 +2,8 @@ package com.sparta.rooibos.deliverer.application.feign;
 
 
 
-import com.sparta.rooibos.user.application.dto.response.UserResponse;
+
+import com.sparta.rooibos.deliverer.application.feign.dto.UserResponse;
 
 import java.util.UUID;
