+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/application/feign/dto/UserResponse.java
@@ -1,9 +1,5 @@
 package com.sparta.rooibos.deliverer.application.feign.dto;
 
-import com.sparta.rooibos.user.domain.entity.Role;
-import com.sparta.rooibos.user.domain.entity.UserRoleStatus;
-
-import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record UserResponse(
@@ -12,13 +8,5 @@ public record UserResponse(
         String email,
         String slackAccount,
         String phone,
-        Role role,
-        boolean hidden,
-        UserRoleStatus status,
-        LocalDateTime createdAt,
-        String createdBy,
-        LocalDateTime updatedAt,
-        String updatedBy,
-        LocalDateTime deletedAt,
-        String deletedBy) {
-}
+        String role
+) {}
