+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/domain/critia/DelivererSearchCriteria.java
@@ -1,0 +1,10 @@
+package com.sparta.rooibos.deliverer.domain.critia;
+
+public record DelivererSearchCriteria(
+        String keyword,
+        String sort,
+        String filter,
+        int page,
+        int size
+) {
+}
