+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/infrastructure/feign/UserClient.java
@@ -1,7 +1,7 @@
 package com.sparta.rooibos.deliverer.infrastructure.feign;
 
 import com.sparta.rooibos.deliverer.application.feign.UserService;
-import com.sparta.rooibos.user.application.dto.response.UserResponse;
+import com.sparta.rooibos.deliverer.application.feign.dto.UserResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
