+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/infrastructure/persistence/DelivererRepositoryCustomImpl.java
@@ -5,6 +5,7 @@ import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.sparta.rooibos.deliverer.application.dto.request.DelivererSearchRequest;
+import com.sparta.rooibos.deliverer.domain.critia.DelivererSearchCriteria;
 import com.sparta.rooibos.deliverer.domain.entity.Deliverer;
 import com.sparta.rooibos.deliverer.domain.entity.QDeliverer;
 import com.sparta.rooibos.deliverer.domain.model.Pagination;
@@ -24,7 +25,7 @@ public class DelivererRepositoryCustomImpl implements DelivererRepositoryCustom 
     private final AuthorizationUtil authorizationUtil;
 
     @Override
-    public Pagination<Deliverer> searchDeliverers(DelivererSearchRequest request) {
+    public Pagination<Deliverer> searchDeliverers(DelivererSearchCriteria request) {
         QDeliverer deliverer = QDeliverer.deliverer;
 
         BooleanBuilder where = authorizationUtil.getPermissionWhereClause(deliverer);
